python/requirements.txt
absl-py==0.12.0
ad3 @ file:///C:/Users/YMZ/Desktop/nativeProject/python/py_PyUserInput/ad3-2.2.1-cp38-cp38-win_amd64.whl
alembic==1.6.5
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-lru==1.0.2
autopep8==1.5.7
Babel==2.9.1
bce-python-sdk==0.8.61
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1605538112425/work
cfgv==3.3.0
chardet @ file:///C:/ci/chardet_1605303225733/work
click==8.0.1
colorama==0.4.4
colorlog==5.0.1
conda==4.9.2
conda-package-handling @ file:///C:/ci/conda-package-handling_1603003327818/work
cryptography @ file:///C:/ci/cryptography_1605526388508/work
cycler==0.10.0
DBUtils==2.0.1
decorator==5.0.7
dill==0.3.4
distlib==0.3.2
easydict==1.9
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
Flask-Migrate==3.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.0
grpcio==1.32.0
h5py==2.10.0
identify==2.2.10
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
menuinst==1.4.16
MouseInfo==0.1.3
multiprocess==0.70.12.2
nanoid==2.0.0
nodeenv==1.6.0
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.2.54
opt-einsum==3.3.0
packaging==21.0
paddle2onnx==0.6
paddlehub==2.1.0
paddlenlp==2.0.5
paddlepaddle==2.0.2
pandas==1.3.0
Pillow==8.2.0
pre-commit==2.13.0
protobuf==3.16.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.7.0
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.10.1
pyflakes==2.3.1
PyGetWindow==0.0.9
PyJWT==2.1.0
pykeyboard==0.1.2
PyMouse==1.0
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
Pyrogram==1.2.9
PyScreeze==0.1.27
PySocks @ file:///C:/ci/pysocks_1605287845585/work
python-dateutil==2.8.1
python-editor==1.0.4
PyTweening==1.0.3
pytz==2021.1
pywin32==227
pyWinhook @ file:///C:/Users/YMZ/Desktop/nativeProject/python/py_PyUserInput/pyWinhook-1.6.2-cp38-cp38-win_amd64.whl
PyYAML==5.4.1
pyzmq==22.1.0
rarfile==4.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.7.2
ruamel-yaml==0.15.87
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
seqeval==1.2.2
shellcheck-py==0.7.2.1
six @ file:///C:/ci/six_1605187374963/work
sklearn==0.0
smmap==4.0.0
SQLAlchemy==1.4.20
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
TgCrypto==1.2.2
threadpoolctl==2.1.0
toml==0.10.2
tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
typing-extensions==3.7.4.3
urllib3 @ file:///tmp/build/80754af9/urllib3_1603305693037/work
virtualenv==20.4.7
visualdl==2.2.0
Werkzeug==2.0.1
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306167264/work
wincertstore==0.2
wrapt==1.12.1
